refactor: implement editing and deleting comments with useOptimistic

+++ components/comment/comment-like-button.tsx
@@ -32,67 +32,53 @@ export const CommentLIkeButton: React.FC<CommentLIkeButtonProps> = ({
   const supabase = createSupabaseBrowserClient()
   const [isPending, startTransition] = useTransition()
   const openAccountModal = useAccountModal((state) => state.openModal)
-
   const router = useRouter()
   const { data: profile } = useUserProfile()
 
-  const handleLikeDebounced = useDebouncedCallback(
-    async (profile: Profiles) => {
-      const { error: addLikeError } = await supabase
-        .from(commentLikesTable)
-        .insert({
-          comment_id: comment.comment_id,
-          user_id: profile.user_id,
-        })
-      if (addLikeError) {
-        toast.error("Failed to like. Please try again.")
-      }
-    },
-    1000, // Adjust the debounce delay (in milliseconds) as needed,
-    { leading: true, trailing: true }
-  )
-
-  const handleRemoveLikeDebounced = useDebouncedCallback(
-    async (profile: Profiles) => {
-      const { error: removeLikeError } = await supabase
-        .from(commentLikesTable)
-        .delete()
-        .match({ comment_id: comment.comment_id, user_id: profile.user_id })
-
-      if (removeLikeError) {
-        toast.error("Failed to remove like. Please try again.")
+  const handleLikeToggleDebounced = useDebouncedCallback(
+    async (profile: Profiles, isLiked: boolean) => {
+      if (isLiked) {
+        const { error: removeLikeError } = await supabase
+          .from(commentLikesTable)
+          .delete()
+          .match({ comment_id: comment.comment_id, user_id: profile.user_id })
+        if (removeLikeError) {
+          toast.error("Failed to remove like. Please try again.")
+        }
+      } else {
+        const { error: addLikeError } = await supabase
+          .from(commentLikesTable)
+          .insert({ comment_id: comment.comment_id, user_id: profile.user_id })
+        if (addLikeError) {
+          toast.error("Failed to like. Please try again.")
+        }
       }
     },
     1000,
     { leading: true, trailing: true }
   )
 
-  const handleLikes = async () => {
-    // checking if we have a user login
+  const handleLikes = () => {
     if (!profile?.user_id) {
       toast.error("Please login to like a comment.")
       openAccountModal()
       return
     }
 
-    if (comment.user_has_liked_comment) {
-      setOptimisticComment({
+    const isLiked = comment.user_has_liked_comment || false
+
+    setOptimisticComment({
+      type: "update",
+      comment: {
         ...comment,
-        likes_count: comment.likes_count - 1,
-        user_has_liked_comment: !comment.user_has_liked_comment,
-      })
+        likes_count: isLiked
+          ? comment.likes_count - 1
+          : comment.likes_count + 1,
+        user_has_liked_comment: !isLiked,
+      },
+    })
 
-      handleRemoveLikeDebounced(profile)
-    } else {
-      setOptimisticComment({
-        ...comment,
-        likes_count: comment.likes_count + 1,
-        user_has_liked_comment: !comment.user_has_liked_comment,
-      })
-
-      handleLikeDebounced(profile)
-    }
-
+    handleLikeToggleDebounced(profile, isLiked)
     router.refresh()
   }
 
@@ -103,7 +89,7 @@ export const CommentLIkeButton: React.FC<CommentLIkeButtonProps> = ({
       >
         <button
           className={cn("group rounded-full")}
-          onClick={() => startTransition(() => handleLikes())}
+          onClick={() => startTransition(handleLikes)}
         >
           <Heart
             className={cn(
